refactor: extract cart business logic into cartService
